Big update from setup on new work computer
Really big update after this has been dormant for a year and a half or
so. Back in true terminal vim after a while with vim emulation in
different IDEs and editors. Added a decent amount of new plugins
tailored around the work I'm currently doing (mainly Go), removed one or
two, changed some settings I previously liked, added some mappings, etc.
A lot of it is based around getting vim to play nice on WSL2 in Windows
Terminal. The state of this is pretty messy but I decided to commit it
now so I could synch the changes to my personal rig. Should be a
cleanup/reorg coming soon.